Refactor the SFX editor. Remove the possibility to put 2 notes on the same beat

+++ tiny-engine/src/commonTest/kotlin/com/github/minigdx/tiny/platform/test/HeadlessPlatform.kt
@@ -96,7 +96,7 @@ class HeadlessPlatform(override val gameOptions: GameOptions, val resources: Map
                 }
             }
 
-            override fun playBuffer(buffer: FloatArray, numberOfSamples: Int) = Unit
+            override fun playBuffer(buffer: FloatArray, numberOfSamples: Long) = Unit
         }
     }
