+++ tiny-engine/src/jsMain/kotlin/com/github/minigdx/tiny/platform/webgl/PicoAudioSoundMananger.kt
@@ -68,14 +68,14 @@ class PicoAudioSoundMananger : SoundManager() {
         return PicoAudioSound(audio, smf)
     }
 
-    override fun playBuffer(buffer: FloatArray, numberOfSamples: Int) {
+    override fun playBuffer(buffer: FloatArray, numberOfSamples: Long) {
         val result = convertBuffer(buffer, numberOfSamples)
         playSfxBuffer(result)
     }
 
-    private fun convertBuffer(buffer: FloatArray, length: Int): Float32Array {
-        val result = Float32Array(length)
-        (0 until length).forEach { index ->
+    private fun convertBuffer(buffer: FloatArray, length: Long): Float32Array {
+        val result = Float32Array(length.toInt())
+        (0 until length.toInt()).forEach { index ->
             val byte = buffer[index]
             result[index] = byte
         }
